25/6/2024 [v0.0.0.7 prototype]
Oops, forgot to separate LanguageDialogue from SideNav!

Changed:
> Language Dialogue is now ACTUALLY an individual component
