+++ src/renderer/src/components/Navigation/SideNav.tsx
@@ -1,16 +1,15 @@
-import { ReactElement, useContext, useState } from 'react'
+import React, { ReactElement } from 'react'
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-ignore
 import { NavLink } from 'react-router-dom'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { IconProp } from '@fortawesome/fontawesome-svg-core'
-import i18next from 'i18next'
+
 import { useTranslation } from 'react-i18next'
 
-import '../Language/LanguageDialogue.css'
 import './SideNav.css'
 
-import { LanguageContext } from '../../context/Client/Language/Language'
+import LanguageDialogue from '../Language/LanguageDialogue'
 
 const SideNav = (): ReactElement => {
 
@@ -18,68 +17,18 @@ const SideNav = (): ReactElement => {
   // @ts-ignore
   const { t } = useTranslation()
 
-  const [openDialogue, setOpenDialogue] = useState(false)
+  const languageDialogueRef = React.useRef(null)
 
-  function changeDialogueState(): void {
-    setOpenDialogue(!openDialogue)
-  }
-
-  // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-  // @ts-ignore
-  const { setLanguage } = useContext(LanguageContext)
-
-  function handleChangeLanguage(newLanguage): void {
-    i18next.changeLanguage(newLanguage).then(() => {
-      setLanguage(newLanguage)
-    })
-    changeDialogueState()
+  function openLanguageDialogue(): void {
+    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+    // @ts-ignore
+    languageDialogueRef.current.openDialogue()
   }
 
   return (
     <>
 
-      {/* background overlay element for dialogue box */}
-      <div
-        className="dialogueOverlay"
-        style={{ display: openDialogue ? 'block' : 'none' }}
-        onClick={changeDialogueState}
-      ></div>
-
-      {/* language change dialogue box */}
-      <div
-        className={`languageDialogue ${openDialogue ? 'dialogueScaleUp' : 'dialogueScaleDown'}`}
-        style={{ display: openDialogue ? 'block' : 'none' }}
-      >
-        <div className={'dialogue'}>
-
-          <p id="languageDialogueText" className="dialogueText">
-            {t('languageDialogue.prompt')}
-          </p>
-
-          <div className={'options'}>
-
-            <div id="englishOption" className="dialogueOption" onClick={() => handleChangeLanguage('en')}>
-              <p id="englishOptionText" className="optionText">
-                English (en)
-              </p>
-            </div>
-
-            <div id="koreanOption" className="dialogueOption" onClick={() => handleChangeLanguage('kr')}>
-              <p id="koreanOptionText" className="optionText">
-                한국어 (ko)
-              </p>
-            </div>
-
-            <div id="cancelOption" className="dialogueOption" onClick={changeDialogueState}>
-              <p id="cancelOptionText" className="optionText">
-                {t('languageDialogue.cancel')}
-              </p>
-            </div>
-
-          </div>
-
-        </div>
-      </div>
+      <LanguageDialogue ref={languageDialogueRef}></LanguageDialogue>
 
       {/* Left Side of Window */}
       <div id="navSide" className="navSideShow">
@@ -219,7 +168,7 @@ const SideNav = (): ReactElement => {
         <div id="navSideBottom" className="navSideSection">
 
           {/* language icon */}
-          <div id="navSideLanguage" className="navSideItem" onClick={changeDialogueState}>
+          <div id="navSideLanguage" className="navSideItem" onClick={openLanguageDialogue}>
             <FontAwesomeIcon
               icon={'fa-solid fa-language' as IconProp}
               id="sideLanguageIcon"
